Add Coverage Tab
Add Coverage tab to show dots on the display instead of printing only
the currently seen aircrafts. This allows one to test their coverage
across the sky with related from current lat/long positions.

Instead of a HashMap for storing ICAO aircrafts, we compute the current
detected aircrafts and compute their lat/long and add them to a Vec.

Signed-off-by: wcampbell <wcampbell1995@gmail.com>