[radar] Add --filter_time to reduce mis-decode
Add --filter_time for removing airplanes after the the specified time.
This is down to 2 seconds in the ICAO document, but I see 10 seconds as
being pretty good for reducing the amount of mis-decodes greatly.

This was previosuly 60 seconds, which was giving very bad lat/long if a
long period of time was between both messages odd/even.

See #21