Add heatmap to coverage
Start the default color of an airplane with one ICAO lat/long passover
to gray, with this increasing to white the more times an airplane passes
over this spot.

This also reduces the accuracy of the Coverage display, but the heatmap
requires this or nothing will ever show as being close.