Refactor multiple types into smaller types
Reduce the size of types to reduce memory usage of application and
library for embeded environs.

struct Altitude::alt has been changed to an Option<u16>. The change from
u32 to u16 was done for size reasons, and a normal operating altitude is
lower then the max u16. The Option was to fix the function returning 0
when it's actually a very invalid altitude.

AC13Field altitude was changed to a u16, this isn't usually a user
facing struct.

In libadsb_common, AirplaneDetails altitude was changed to u16, heading
was changed to f32. num_messages was changed to u32.