Add --retry-tcp for retry for dump1090 connection
Add --retry-tcp for connecting to a dump1090 instance if that service
goes down.

The init_tcp_reader is used for the initial connection, which adds a tui
display for waiting for the connection which is nice. This same tui
display is shown when the server goes down and we try again.

QuitReason has been added for some better exit reasons